+++ src/Controller/LeagueController.php
@@ -3,6 +3,7 @@
 namespace App\Controller;
 
 use App\Repository\LeagueRepository;
+use App\Repository\MottoRepository;
 use App\Repository\StatsRepository;
 use App\Repository\UserRepository;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
@@ -14,27 +15,38 @@ class LeagueController extends AbstractController
     /**
      * @Route("/league/{league_name}", name="app_league")
      */
-    public function index(Request $request, LeagueRepository $leagueRepository, UserRepository $userRepository, StatsRepository $statsRepository)
-    {
+    public function index(
+        Request $request,
+        LeagueRepository $leagueRepository,
+        UserRepository $userRepository,
+        StatsRepository $statsRepository,
+        MottoRepository $mottoRepository
+    ) {
+        // $currentUser = $this->getUser()->getUsername();
         $league = $request->get('league_name');
         // get league ID
         $leagueOBJ = $leagueRepository->findOneBy(
             ['leagueName' => $league]
         );
-        $leaugeID = $leagueOBJ->getId();
+        $leagueID = $leagueOBJ->getId();
 
         // get all users in that league
-        $users = $userRepository->findBy(['league' => $leaugeID]);
+        $users = $userRepository->findBy(['league' => $leagueID]);
         
         // tally up player scores
         foreach ($users as $user) {
             $total = 0;
             $playerStats = $statsRepository->findBy(['player' => $user->getId()]);
+            $playerMotto = $mottoRepository->findMotto($user->getId())['motto'];
             foreach ($playerStats as $stat) {
                 $total += $stat->getPoints();
             }
-            // add total property to user object so I can call it in the template
+            /*
+            add `total` and `motto` properties to user object
+            so I can call them from the league template
+            */
             $user->{'total'} = $total;
+            $user->{'motto'} = $playerMotto;
         };
         // sorting users from highest to lowest score
         usort(
@@ -49,6 +61,7 @@ class LeagueController extends AbstractController
         return $this->render(
             'league/index.html.twig',
             [
+                // 'currentUser' => $currentUser,
                 'league' => $league,
                 'users' => $users,
             ]